Make NO_CHAPTER_INDEX_ENTRY unsigned.
Pair:  sclafani

.../chapter-index.h
.../chapter-index.c
.../sparse-cache.h
.../sparse-cache.c
.../volume.h
.../volume.c
.../index.c
.../ChapterIndex_t1.c
Changed NO_CHAPTER_INDEX_ENTRY from -1 to U16_MAX to allow the
 record_page_number returned by a chapter index search to be
 unsigned. Also trimmed those fields to a u16, since the record
  page number within a chapter can't be higher than 255.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158075]

+++ src/c++/uds/src/tests/ChapterIndex_t1.c
@@ -190,9 +190,8 @@ static void basicChapterTest(void)
     for (page = 0; page < g->index_pages_per_chapter; page++) {
       if ((cip[page].lowest_list_number <= deltaListNumber)
           && (deltaListNumber <= cip[page].highest_list_number)) {
-        int entry;
-        UDS_ASSERT_SUCCESS(search_chapter_index_page(&cip[page], g, &names[i],
-                                                     &entry));
+        u16 entry;
+        UDS_ASSERT_SUCCESS(search_chapter_index_page(&cip[page], g, &names[i], &entry));
         CU_ASSERT_EQUAL(entry, generatePageNumber(g, i));
         inChapter = true;
       }
